Adding data alignment and double precision to RMN ops

+++ rocAL/rocAL/source/meta_data/meta_node_resize_mirror_normalize.cpp
@@ -86,8 +86,8 @@ void ResizeMirrorNormalizeMetaNode::update_parameters(MetaDataBatch *input_meta_
         {            
             if(_mirror_val[i] == 1)
             {
-                float one_by_width_coeff = 1 / float(_dst_width_val[i]);
-                float l = 1 - coords_buf[j].r - one_by_width_coeff;
+                double one_by_width_coeff = 1 / double(_dst_width_val[i]);
+                double l = 1 - coords_buf[j].r - one_by_width_coeff;
                 coords_buf[j].r = 1 - coords_buf[j].l - one_by_width_coeff;
                 coords_buf[j].l = l; 
             }
